+++ src/solvers/ssm/SimInf_solver_ssm.c
@@ -5,7 +5,7 @@
  * Copyright (C) 2015 Pavol Bauer
  * Copyright (C) 2017 -- 2019 Robin Eriksson
  * Copyright (C) 2015 -- 2019 Stefan Engblom
- * Copyright (C) 2015 -- 2021 Stefan Widgren
+ * Copyright (C) 2015 -- 2022 Stefan Widgren
  *
  * SimInf is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -73,6 +73,8 @@ static int SimInf_solver_ssm(
                     m.sum_t_rate[node] += rate;
                     if (!R_FINITE(rate) || rate < 0.0) {
                         SimInf_print_status(m.Nc, &m.u[node * m.Nc],
+                                            m.Nd, &m.v[node * m.Nd],
+                                            m.Nld, &m.ldata[node * m.Nld],
                                             m.Ni + node, m.tt, rate, j);
                         m.error = SIMINF_ERR_INVALID_RATE;
                     }
@@ -158,6 +160,8 @@ static int SimInf_solver_ssm(
                             m.u[node * m.Nc + m.irS[j]] += m.prS[j];
                             if (m.u[node * m.Nc + m.irS[j]] < 0) {
                                 SimInf_print_status(m.Nc, &m.u[node * m.Nc],
+                                                    m.Nd, &m.v[node * m.Nd],
+                                                    m.Nld, &m.ldata[node * m.Nld],
                                                     m.Ni + node, m.t_time[node],
                                                     0, tr);
                                 m.error = SIMINF_ERR_NEGATIVE_STATE;
@@ -177,6 +181,8 @@ static int SimInf_solver_ssm(
                             delta += rate - old;
                             if (!R_FINITE(rate) || rate < 0.0) {
                                 SimInf_print_status(m.Nc, &m.u[node * m.Nc],
+                                                    m.Nd, &m.v[node * m.Nd],
+                                                    m.Nld, &m.ldata[node * m.Nld],
                                                     m.Ni + node, m.t_time[node],
                                                     rate, m.irG[j]);
                                 m.error = SIMINF_ERR_INVALID_RATE;
@@ -244,6 +250,8 @@ static int SimInf_solver_ssm(
                             delta += rate - old;
                             if (!R_FINITE(rate) || rate < 0.0) {
                                 SimInf_print_status(m.Nc, &m.u[node * m.Nc],
+                                                    m.Nd, &m.v[node * m.Nd],
+                                                    m.Nld, &m.ldata[node * m.Nld],
                                                     m.Ni + node, m.tt, rate, j);
                                 m.error = SIMINF_ERR_INVALID_RATE;
                             }